print for testing

+++ dl_framework/data.py
@@ -15,10 +15,10 @@ def do_normalisation(x, norm):
     :param x        Object to be normalized
     :param norm     Pandas Dataframe which includes the normalisation factors
     """
-    train_mean_c0 = torch.tensor(norm["train_mean_c0"].values[0]).float()
-    train_std_c0 = torch.tensor(norm["train_std_c0"].values[0]).float()
-    train_mean_c1 = torch.tensor(norm["train_mean_c1"].values[0]).float()
-    train_std_c1 = torch.tensor(norm["train_std_c1"].values[0]).float()
+    train_mean_c0 = torch.tensor(norm["train_mean_c0"].values[0]).double()
+    train_std_c0 = torch.tensor(norm["train_std_c0"].values[0]).double()
+    train_mean_c1 = torch.tensor(norm["train_mean_c1"].values[0]).double()
+    train_std_c1 = torch.tensor(norm["train_std_c1"].values[0]).double()
     x[:, 0] = normalize(x[:, 0], train_mean_c0, train_std_c0)
     x[:, 1] = normalize(x[:, 1], train_mean_c1, train_std_c1)
     assert not torch.isinf(x).any()
